introduced RTL::BitFieldIndex module (refs: #45)
